src/requirement.txt
Django==2.0.7
django-crispy-forms==1.9.2
django-tinymce==3.1.0
Pillow==7.2.0
pytz==2020.1
